contrib/labstack/echo.v4: refactor the appsec test into echotrace_test.go
