+++ contrib/labstack/echo.v4/echotrace_test.go
@@ -7,16 +7,20 @@ package echo
 
 import (
 	"errors"
+	"io/ioutil"
 	"net/http"
 	"net/http/httptest"
+	"strings"
 	"testing"
 
 	"gopkg.in/DataDog/dd-trace-go.v1/ddtrace/ext"
 	"gopkg.in/DataDog/dd-trace-go.v1/ddtrace/mocktracer"
 	"gopkg.in/DataDog/dd-trace-go.v1/ddtrace/tracer"
+	"gopkg.in/DataDog/dd-trace-go.v1/internal/appsec"
 
 	"github.com/labstack/echo/v4"
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
 
 func TestChildSpan(t *testing.T) {
@@ -229,4 +233,48 @@ func TestGetSpanNotInstrumented(t *testing.T) {
 	router.ServeHTTP(w, r)
 	assert.True(called)
 	assert.False(traced)
+}
+
+func TestAppSec(t *testing.T) {
+	// Start the tracer along with the fake agent HTTP server
+	mt := mocktracer.Start()
+	defer mt.Stop()
+
+	appsec.Start()
+	defer appsec.Stop()
+
+	if !appsec.Enabled() {
+		t.Skip("appsec disabled")
+	}
+
+	// Start and trace an HTTP server
+	e := echo.New()
+	e.Use(Middleware())
+
+	e.POST("/*tmp", func(c echo.Context) error {
+		return c.String(200, "Hello World!\n")
+	})
+	srv := httptest.NewServer(e)
+	defer srv.Close()
+
+	// Send an LFI attack
+	req, err := http.NewRequest("POST", srv.URL+"/../../../secret.txt", nil)
+	if err != nil {
+		panic(err)
+	}
+	res, err := srv.Client().Do(req)
+	require.NoError(t, err)
+
+	// Check that the handler was properly called
+	b, err := ioutil.ReadAll(res.Body)
+	require.NoError(t, err)
+	require.Equal(t, "Hello World!\n", string(b))
+
+	finished := mt.FinishedSpans()
+	require.Len(t, finished, 1)
+
+	// The request should have the LFI attack attempt event (appsec rule id crs-930-100).
+	event := finished[0].Tag("_dd.appsec.json")
+	require.NotNil(t, event)
+	require.True(t, strings.Contains(event.(string), "crs-930-100"))
 }
